Prevent system_server shutdown when trying to format external storage
The SystemUI does not provied path information on which volume that should
be formatted. So when the user selects "Blank CD Card" an intent is sent to
ExternalMediaFormatActivity and forwarded to ExternalStorageFormatter then
actual volume that is unformatted is not included in the information sent
in the Intent.
This causes the ExternalStorageFormatter to go through Environment to find
a Volume but the actual volume found that way is not correct.

To solve the issue look up the correct StorageVolumeand reuse the existing
EXTRA_STORAGE_VOLUME extra and send it as part of the Intent.

Change-Id: Ib360a9ce071ebe95fc3ca847ed68b6db05928a42

Conflicts:
	core/java/com/android/internal/app/ExternalMediaFormatActivity.java
	packages/SystemUI/src/com/android/systemui/usb/StorageNotification.java